+++ ignite/pkg/xast/global_test.go
@@ -48,6 +48,30 @@ var myIntVar int = 42
 `,
 		},
 		{
+			name: "Insert global int var without type",
+			args: args{
+				fileContent: `package main
+
+import (
+	"fmt"
+)
+
+`,
+				globalType: GlobalTypeVar,
+				globals: []GlobalOptions{
+					WithGlobal("myIntVar", "", "42"),
+				},
+			},
+			want: `package main
+
+import (
+	"fmt"
+)
+
+var myIntVar = 42
+`,
+		},
+		{
 			name: "Insert global int const",
 			args: args{
 				fileContent: `package main
@@ -530,6 +554,110 @@ type MyStruct struct {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			got, err := ModifyStruct(tt.args.fileContent, tt.args.structName, tt.args.options...)
+			if tt.err != nil {
+				require.Error(t, err)
+				require.Equal(t, tt.err.Error(), err.Error())
+				return
+			}
+			require.NoError(t, err)
+			require.Equal(t, tt.want, got)
+		})
+	}
+}
+
+func TestModifyGlobalArrayVar(t *testing.T) {
+	type args struct {
+		fileContent string
+		globalName  string
+		options     []GlobalArrayOpts
+	}
+	tests := []struct {
+		name string
+		args args
+		want string
+		err  error
+	}{
+		{
+			name: "Add field to custom variable array",
+			args: args{
+				fileContent: `package app
+var (
+	moduleAccPerms = []*authmodulev1.ModuleAccountPermission{
+		{Account: nft.ModuleName},
+		{Account: ibctransfertypes.ModuleName, Permissions: []string{authtypes.Minter, authtypes.Burner}},
+	}
+)
+`,
+				globalName: "moduleAccPerms",
+				options:    []GlobalArrayOpts{AppendGlobalArrayValue("{Account: icatypes.ModuleName}")},
+			},
+			want: `package app
+
+var (
+	moduleAccPerms = []*authmodulev1.ModuleAccountPermission{
+		{Account: nft.ModuleName},
+		{Account: ibctransfertypes.ModuleName, Permissions: []string{authtypes.Minter, authtypes.Burner}},
+		{Account: icatypes.ModuleName},
+	}
+)
+`,
+		},
+		{
+			name: "Add field to string variable array",
+			args: args{
+				fileContent: `package app
+
+var (
+	blockAccAddrs = []string{
+		authtypes.FeeCollectorName,
+		distrtypes.ModuleName,
+		minttypes.ModuleName,
+		stakingtypes.BondedPoolName,
+		stakingtypes.NotBondedPoolName,
+	}
+)
+`,
+				globalName: "blockAccAddrs",
+				options:    []GlobalArrayOpts{AppendGlobalArrayValue("nft.ModuleName")},
+			},
+			want: `package app
+
+var (
+	blockAccAddrs = []string{
+		authtypes.FeeCollectorName,
+		distrtypes.ModuleName,
+		minttypes.ModuleName,
+		stakingtypes.BondedPoolName,
+		stakingtypes.NotBondedPoolName,
+		nft.ModuleName,
+	}
+)
+`,
+		},
+		{
+			name: "name not found",
+			args: args{
+				fileContent: `package app
+
+var (
+	blockAccAddrs = []string{
+		authtypes.FeeCollectorName,
+		distrtypes.ModuleName,
+		minttypes.ModuleName,
+		stakingtypes.BondedPoolName,
+		stakingtypes.NotBondedPoolName,
+	}
+)
+`,
+				globalName: "notFound",
+				options:    []GlobalArrayOpts{AppendGlobalArrayValue("nft.ModuleName")},
+			},
+			err: errors.New("global array \"notFound\" not found in file content"),
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := ModifyGlobalArrayVar(tt.args.fileContent, tt.args.globalName, tt.args.options...)
 			if tt.err != nil {
 				require.Error(t, err)
 				require.Equal(t, tt.err.Error(), err.Error())