+++ changelog.md
@@ -2,11 +2,16 @@
 
 ## Unreleased
 
+### Features
+
+- [#4790](https://github.com/ignite/cli/pull/4790) Remove global vars and struct placeholders.
+
 ## [`v29.3.1`](https://github.com/ignite/cli/releases/tag/v29.3.1)
 
 ### Fixes
 
 - [#4793](https://github.com/ignite/cli/pull/4793) Use latest `bytedance/sonic` version to support Go 1.25.
+
 
 ## [`v29.3.0`](https://github.com/ignite/cli/releases/tag/v29.3.0)